splitout concurrency tests
